+++ lib/sidekiq/web.rb
@@ -39,7 +39,7 @@ module Sidekiq
     end
 
     get "/workers" do
-      erb :index
+      erb :workers
     end
 
     get "/queues" do